refactor(workspace): align controller, service, and routes with new permission system

+++ server/src/modules/Workspace/workspace.controller.ts
@@ -3,6 +3,14 @@ import { AppError } from "../../utils/appError.js";
 import { WorkspaceService } from "./workspace.service.js";
 
 export default class WorkspaceController {
+  // GET: /workspaces
+  static async getWorkspaces(req: Request, res: Response) {
+    const userId = req.user.id;
+
+    const workspaces = await WorkspaceService.getUserWorkspaces(userId);
+
+    res.status(200).json(workspaces);
+  }
   // POST: /workspaces
   static async createWorkspace(req: Request, res: Response) {
     const userId = req.user.id;
@@ -14,57 +22,49 @@ export default class WorkspaceController {
     res.status(201).json(workspace);
   }
 
-  // GET: /workspaces
-  static async getWorkspaces(req: Request, res: Response) {
-    const userId = req.user.id;
-
-    const workspaces = await WorkspaceService.getUserWorkspaces(userId);
-
-    res.status(200).json(workspaces);
-  }
-
   // GET: /workspaces/:workspaceId
-  static async getWorkspaceById(req: Request, res: Response) {
-    const userId = req.user.id;
+  static async getWorkspace(req: Request, res: Response) {
     const { workspaceId } = req.params;
-
     if (!workspaceId) throw new AppError("Workspace id not provided", 400);
 
-    const workspace = await WorkspaceService.getWorkspaceById(
+    const workspace = await WorkspaceService.getWorkspace(
       workspaceId as string,
-      userId,
     );
     if (!workspace) throw new AppError("Not found workspace with id", 404);
 
     res.status(200).json(workspace);
   }
 
+  // PATCH: /workspaces/:workspaceId
+  static async editWorkspace(req: Request, res: Response) {
+    const { workspaceId } = req.params;
+    const { name } = req.body;
+    if (!workspaceId) throw new AppError("Workspace id not provided", 400);
+
+    const workspace = await WorkspaceService.editWorkspace(
+      workspaceId as string,
+      name,
+    );
+
+    if (!workspace) throw new AppError("Forbidden", 500);
+
+    res.status(200).json(workspace);
+  }
+
   // DELETE: /workspaces/:workspaceId
   static async deleteWorkspace(req: Request, res: Response) {
-    const userId = req.user.id;
     const { workspaceId } = req.params;
 
     if (!workspaceId) throw new AppError("Workspace id not provided", 400);
 
     const deletedWorkspace = await WorkspaceService.deleteWorkspace(
       workspaceId as string,
-      userId,
     );
 
     res.status(200).json({ message: deletedWorkspace.message });
   }
 
-  // PATCH: /workspaces/:workspaceId/members
-  static async addMember(req: Request, res: Response) {
-    const workspaceId = req.params.workspaceId as string;
-    const { userId, role } = req.body;
-
-    if (!workspaceId) throw new AppError("Workspace id not provided", 400);
-
-    const member = await WorkspaceService.addMember(workspaceId, userId, role);
-
-    res.status(201).json(member);
-  }
+  // ========================== WORKSPACE MEMBER ==========================
 
   // GET /workspaces/:workspaceId/members
   static async getMembers(req: Request, res: Response) {
@@ -75,10 +75,24 @@ export default class WorkspaceController {
 
     res.status(200).json(members);
   }
+
+  // PATCH: /workspaces/:workspaceId/members
+  static async addMember(req: Request, res: Response) {
+    const workspaceId = req.params.workspaceId as string;
+    const { userId, role } = req.body;
+    if (!workspaceId) throw new AppError("Workspace id not provided", 400);
+
+    const member = await WorkspaceService.addMember(workspaceId, userId, role);
+
+    res.status(201).json(member);
+  }
+
   // PATCH /workspaces/:workspaceId/members/:memberId
   static async editMember(req: Request, res: Response) {
     const { workspaceId, memberId } = req.params;
     const { role } = req.body;
+
+    if(!workspaceId || !memberId) throw new AppError("Missing params", 400)
 
     const member = await WorkspaceService.editMember(
       workspaceId as string,
@@ -92,20 +106,33 @@ export default class WorkspaceController {
   // DELETE /workspaces/:workspaceId/members/:memberId
   static async deleteMember(req: Request, res: Response) {
     const { workspaceId, memberId } = req.params;
-    const member = await WorkspaceService.deleteMember(
+    if(!workspaceId || !memberId) throw new AppError("Missing params", 400)
+
+    await WorkspaceService.deleteMember(
       workspaceId as string,
       memberId as string,
     );
 
-    res.status(200).json({ message: "Delete success" });
+    res.status(200).json({ message: "Member removed" });
   }
 
-    // GET /workspaces/:workspaceId/boards
-    static async getBoards(req: Request, res: Response) {
-      const workspaceId = req.params.workspaceId as string;
-  
-      const boards = await WorkspaceService.getBoards(workspaceId);
-  
-      res.status(200).json(boards);
-    }
+  // GET /workspaces/:workspaceId/boards
+  static async getBoards(req: Request, res: Response) {
+    const workspaceId = req.params.workspaceId as string;
+
+    const boards = await WorkspaceService.getBoards(workspaceId);
+
+    res.status(200).json(boards);
+  }
+
+  // POST /workspaces/:workspaceId/boards
+  static async createBoard(req: Request, res: Response) {
+    const workspaceId = req.params.workspaceId as string
+    const { title, visibility, background } = req.body
+    const userId = req.user.id
+    
+    const board = await WorkspaceService.createBoard(workspaceId, title, visibility, background, userId)
+
+    res.status(200).json()
+  }
 }
